src/processing_consumer/requirements.txt
dapr-ext-grpc
cloudevents